requirements.txt
Django>=4.2.1
django_debug_toolbar
djangorestframework
djangorestframework-simplejwt
psycopg2
psycopg2-binary
dj-database-url
python-decouple
django-cors-headers
whitenoise
django-filter
pytest
pytest-django
factory-boy
pytest-mock
Faker
django-cloudinary-storage
cloudinary
Pillow
drf-spectacular
python-dotenv



